Rename View->views and ViewModel -> view_model
