+++ lib/views/mycourses_screen.dart
@@ -1,18 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:gocast_mobile/View/utils/course_card_view.dart';
+import 'package:gocast_mobile/views/utils/course_card_view.dart';
 import 'utils/course_screen.dart';
 import 'utils/constants.dart';
-class PublicCourses extends StatelessWidget {
-  const PublicCourses({super.key});
+
+// MyCourses.dart
+class MyCourses extends StatelessWidget {
+  const MyCourses({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const CoursesScreen(
-      title: 'Public Courses',
+      title: 'My Courses',
       courseCards: [
         CourseCard(
           title: 'PSY101',
-          subtitle: 'Public Psychology Course',
+          subtitle: 'Introduction to Psychology',
           path: courseImage1,
         ),
         // Add more courses as needed
