+++ lib/views/pinnedcourses_screen.dart
@@ -1,21 +1,20 @@
-// DownloadsScreen.dart
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'utils/courselist_screen.dart';
-import 'package:gocast_mobile/View/utils/video_card_view.dart';
+import 'package:gocast_mobile/views/utils/video_card_view.dart';
 
-class DownloadsScreen extends ConsumerWidget {
-  const DownloadsScreen({super.key});
+class PinnedCourses extends ConsumerWidget {
+  const PinnedCourses({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     return const CourseListScreen(
-      title: 'Downloads',
+      title: 'Pinned',
       videoCards: [
         VideoCard(
-          imageName: 'assets/images/course1.png',
-          title: 'Lineare Algebra für Informatik [MA0901]',
-          date: 'July 24, 2019',
+          imageName: 'assets/images/course2.png',
+          title: 'Computer Science [CS202]',
+          date: 'July 23, 2019',
           duration: '02:00:00',
         ),
         // Add more VideoCard widgets as needed
