+++ lib/views/publiccourses_screen.dart
@@ -1,19 +1,19 @@
 import 'package:flutter/material.dart';
-import 'package:gocast_mobile/View/utils/course_card_view.dart';
+import 'package:gocast_mobile/views/utils/course_card_view.dart';
 import 'utils/course_screen.dart';
 import 'utils/constants.dart';
-// MyCourses.dart
-class MyCourses extends StatelessWidget {
-  const MyCourses({super.key});
+
+class PublicCourses extends StatelessWidget {
+  const PublicCourses({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const CoursesScreen(
-      title: 'My Courses',
+      title: 'Public Courses',
       courseCards: [
         CourseCard(
           title: 'PSY101',
-          subtitle: 'Introduction to Psychology',
+          subtitle: 'Public Psychology Course',
           path: courseImage1,
         ),
         // Add more courses as needed
