+++ lib/views/utils/course_card_view.dart
