+++ lib/views/utils/custom_bottom_nav_bar.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../courseoverview_screen.dart';
 import '../download_screen.dart';
-import 'package:gocast_mobile/View/pinnedcourses_screen.dart';
-import 'package:gocast_mobile/View/notifications_screen.dart';
+import 'package:gocast_mobile/views/pinnedcourses_screen.dart';
+import 'package:gocast_mobile/views/notifications_screen.dart';
 
 // Assuming currentIndexProvider is defined in a global scope file:
 // final currentIndexProvider = StateProvider<int>((ref) => 0);
@@ -25,25 +25,25 @@ class CustomBottomNavBar extends ConsumerWidget {
         case 0:
           Navigator.of(context).pushAndRemoveUntil(
             MaterialPageRoute(builder: (context) => const CourseOverview()),
-                (Route<dynamic> route) => false,
+            (Route<dynamic> route) => false,
           );
           break;
         case 1:
           Navigator.of(context).pushAndRemoveUntil(
             MaterialPageRoute(builder: (context) => const DownloadsScreen()),
-                (Route<dynamic> route) => false,
+            (Route<dynamic> route) => false,
           );
           break;
         case 2:
           Navigator.of(context).pushAndRemoveUntil(
             MaterialPageRoute(builder: (context) => const PinnedCourses()),
-                (Route<dynamic> route) => false,
+            (Route<dynamic> route) => false,
           );
           break;
         case 3:
           Navigator.of(context).pushAndRemoveUntil(
-            MaterialPageRoute(builder: (context) =>  NotificationsScreen()),
-                (Route<dynamic> route) => false,
+            MaterialPageRoute(builder: (context) => NotificationsScreen()),
+            (Route<dynamic> route) => false,
           );
           break;
         default:
@@ -60,15 +60,17 @@ class CustomBottomNavBar extends ConsumerWidget {
           label: 'Home',
         ),
         BottomNavigationBarItem(
-          icon: Icon(Icons.file_download, color:  _getColorForIcon(1, currentIndex)),
+          icon: Icon(Icons.file_download,
+              color: _getColorForIcon(1, currentIndex)),
           label: 'Downloads',
         ),
         BottomNavigationBarItem(
-          icon: Icon(Icons.push_pin, color:  _getColorForIcon(2, currentIndex)),
+          icon: Icon(Icons.push_pin, color: _getColorForIcon(2, currentIndex)),
           label: 'Pinned',
         ),
         BottomNavigationBarItem(
-          icon: Icon(Icons.notifications, color:  _getColorForIcon(3, currentIndex)),
+          icon: Icon(Icons.notifications,
+              color: _getColorForIcon(3, currentIndex)),
           label: 'Notifications',
         ),
       ],
