+++ lib/views/utils/video_card_view.dart
